+++ fractal_server/app/routes/auth/login.py
@@ -4,7 +4,6 @@ Definition of `/auth/{login,logout}/`, `/auth/token/{login/logout}` routes.
 
 from fastapi import APIRouter
 
-from . import cookie_backend
 from . import fastapi_users
 from . import token_backend
 
@@ -14,9 +13,6 @@ router_login = APIRouter()
 router_login.include_router(
     fastapi_users.get_auth_router(token_backend),
     prefix="/token",
-)
-router_login.include_router(
-    fastapi_users.get_auth_router(cookie_backend),
 )